refactor

+++ utils/helper.js
@@ -10,21 +10,21 @@ export function createMenuItem(menuItemLink, label, key, icon, children) {
 }
 
 export const calculateValue = (registryDayDatas, key, reduce) =>
-  registryDayDatas && parseInt(registryDayDatas[0][key]) / reduce;
+  registryDayDatas && parseFloat(registryDayDatas[0][key]) / reduce;
 
 export const calculatePercentChange = (registryDayDatas, key) => {
-  if (registryDayDatas && parseInt(registryDayDatas[1][key]) === 0) {
+  if (registryDayDatas && parseFloat(registryDayDatas[1][key]) === 0) {
     return {
-      value: parseInt(registryDayDatas[0][key]) * 100,
+      value: parseFloat(registryDayDatas[0][key]) * 100,
       isIncrease: true,
     };
   }
   const value =
     registryDayDatas &&
     (
-      ((parseInt(registryDayDatas[0][key]) -
-        parseInt(registryDayDatas[1][key])) /
-        parseInt(registryDayDatas[1][key])) *
+      ((parseFloat(registryDayDatas[0][key]) -
+        parseFloat(registryDayDatas[1][key])) /
+        parseFloat(registryDayDatas[1][key])) *
       100
     ).toFixed(2);
   return { value: Math.abs(value), isIncrease: value > 0 };
